Handle IJulia via extension if possible, but use Base.require
